feat: implement cascading category reassignment on deletion and zero out product inventory with stock journals upon deletion

+++ backend/src/features/products/controllers/product.controller.ts
@@ -98,13 +98,17 @@ export class ProductController {
   };
 
   public deleteProduct = async (
-    req: Request,
+    req: AuthRequest,
     res: Response,
     next: NextFunction,
   ) => {
     try {
+      const authReq = req as AuthRequest;
+      if (!authReq.user) throw new Error("User unauthorized");
+
       const data = await this.productService.deleteProduct(
         req.params.id as string,
+        authReq.user,
       );
       sendSuccess(res, data, "Delete product successfully");
     } catch (error) {
